[4298] Change default directory, fix prefix length output
Default directory was [prefix]/var/kea/var, rather than
[prefix]/var/kea. Corrected this in code and doc.

V6 Prefix length needed a static cast to int to ensure
numeric rather than character interpretation.

+++ src/hooks/dhcp/legal_log/lease6_callouts.cc
@@ -98,7 +98,8 @@ std::string genLease6Entry(const Pkt6Ptr& query, const Lease6Ptr& lease, const b
     if (lease->type_  != Lease::TYPE_PD) {
         stream <<  "Address:" << lease->addr_;
     } else {
-        stream <<  "Prefix:" << lease->addr_ << "/" << lease->prefixlen_;
+        stream <<  "Prefix:" << lease->addr_ << "/"
+               << static_cast<int>(lease->prefixlen_);
     }
 
     stream << " has been " << (renewal ? "renewed" : "assigned");
